+++ template/internal/ast/ast.go
@@ -321,10 +321,27 @@ func quoteStringIfNeeded(input string) string {
 	if SimpleStringValue.MatchString(input) {
 		return input
 	} else {
-		if strings.ContainsRune(input, '\'') {
-			return "\"" + input + "\""
-		} else {
-			return "'" + input + "'"
-		}
+		return quoteString(input)
 	}
+}
+
+func quoteString(str string) string {
+	if strings.ContainsRune(str, '\'') {
+		return "\"" + str + "\""
+	} else {
+		return "'" + str + "'"
+	}
+}
+
+func isQuoted(str string) bool {
+	if len(str) < 2 {
+		return false
+	}
+	if str[0] == '\'' && str[len(str)-1] == '\'' {
+		return true
+	}
+	if str[0] == '"' && str[len(str)-1] == '"' {
+		return true
+	}
+	return false
 }